[USH 135

+++ database/database.php
@@ -1412,12 +1412,22 @@ public function acceptRentalRequest($request_id) {
     }
 
     public function hardDeleteUnit($space_id) {
+        // Only allow deletion if all invoices for this unit are kicked or there are no invoices
+        $activeInvoice = $this->runQuery(
+            "SELECT 1 FROM invoice WHERE Space_ID = ? AND Flow_Status != 'kicked' LIMIT 1",
+            [$space_id]
+        );
+        if ($activeInvoice) {
+            // There is at least one active (not kicked) invoice, do not delete
+            return false;
+        }
         $this->pdo->beginTransaction();
         try {
             $this->executeStatement("DELETE FROM spaceavailability WHERE Space_ID = ?", [$space_id]);
             $this->executeStatement("DELETE FROM clientspace WHERE Space_ID = ?", [$space_id]);
             $this->executeStatement("DELETE FROM rentalrequest WHERE Space_ID = ?", [$space_id]);
             $this->executeStatement("DELETE FROM maintenancerequest WHERE Space_ID = ?", [$space_id]);
+            $this->executeStatement("DELETE FROM invoice WHERE Space_ID = ?", [$space_id]);
             $this->executeStatement("DELETE FROM space WHERE Space_ID = ?", [$space_id]);
             $this->pdo->commit();
             return true;
